Crate flow_impl

flow_impl is a trait that flow function implementations must implement in order to be executed as part of a flow.

Constants

DONT_RUN_AGAIN
Use DONT_RUN_AGAIN to indicate that a function should not be executed more times

RUN_AGAIN
Use RUN_AGAIN to indicate that a function can be executed more times

Traits

Implementation
An implementation runs with an array of inputs and returns a value (or null) and a bool indicating if it should be ran again.

Type Definitions

RunAgain
Implementations should return a value of type RunAgain to indicate if it should be executed more times in the future.